Add files via upload
Here are the files for the menu implementation. I sort of ditched the polymorphism idea. I left the virtual functions which are still overridden in the child classes, but I figured it would be redundant to use dynamic casting since I'm only ever needing to create two objects. Let me know if there is anything you see wrong or have questions about.