Add files via upload
These are the Car and Obstacles classes. I wrote and sent the main I used to test my code and demonstrate how to use the classes. I'm gonna get on working on modifying the current main with these classes. I haven't yet implemented sound yet, so at the moment when the carcrash function is called it just prints something to the console. I also have an example of how to delete obstacles on the screen after crashing in the main.